refactor(VerifyEmailPane): Convert to TypeScript
